Fix up removal cases
